refeactored order service. new routes controller and service to handle routes
